+++ bindings/nodejs/native/src/classes/client/api.rs
@@ -9,6 +9,7 @@ use crate::classes::client::dto::{AddressBalanceDto, MessageWrapper, OutputMetad
 use iota_client::{
     bee_message::prelude::{Address, MessageBuilder, MessageId, Parents, TransactionId, UtxoInput},
     bee_rest_api::types::dtos::{AddressDto, MessageDto as BeeMessageDto, OutputDto as BeeOutput},
+    common::packable::Packable,
     AddressOutputsOptions, ClientMiner, Seed,
 };
 use neon::prelude::*;
@@ -172,7 +173,7 @@ impl Task for ClientTask {
                     bech32_hrp,
                     include_internal,
                 } => {
-                    let mut getter = client.get_addresses(&seed);
+                    let mut getter = client.get_addresses(seed);
                     if let Some(account_index) = account_index {
                         getter = getter.with_account_index(*account_index);
                     }
@@ -245,16 +246,18 @@ impl Task for ClientTask {
                     serde_json::to_string(&tips)?
                 }
                 Api::PostMessage(message) => {
-                    let parent_msg_ids = match message.parents.as_ref() {
+                    let mut parent_msg_ids = match message.parents.as_ref() {
                         Some(parents) => {
                             let mut parent_ids = Vec::new();
                             for msg_id in parents {
-                                parent_ids.push(MessageId::from_str(&msg_id)?)
+                                parent_ids.push(MessageId::from_str(msg_id)?)
                             }
                             parent_ids
                         }
                         None => client.get_tips().await?,
                     };
+                    parent_msg_ids.sort_unstable_by_key(|a| a.pack_new());
+                    parent_msg_ids.dedup();
                     let network_id = client.get_network_id().await?;
                     let nonce_provider = client.get_pow_provider().await;
                     let message = MessageBuilder::<ClientMiner>::new()
@@ -271,19 +274,19 @@ impl Task for ClientTask {
                     serde_json::to_string(&messages)?
                 }
                 Api::GetMessage(id) => {
-                    let message = client.get_message().data(&id).await?;
+                    let message = client.get_message().data(id).await?;
                     serde_json::to_string(&MessageWrapper {
                         message_id: message.id().0,
                         message: BeeMessageDto::from(&message),
                     })?
                 }
                 Api::GetMessageMetadata(id) => {
-                    let metadata = client.get_message().metadata(&id).await?;
+                    let metadata = client.get_message().metadata(id).await?;
                     serde_json::to_string(&metadata)?
                 }
-                Api::GetRawMessage(id) => client.get_message().raw(&id).await?,
+                Api::GetRawMessage(id) => client.get_message().raw(id).await?,
                 Api::GetMessageChildren(id) => {
-                    let messages = client.get_message().children(&id).await?;
+                    let messages = client.get_message().children(id).await?;
                     serde_json::to_string(&messages)?
                 }
                 Api::GetOutput(id) => {