+++ examples/05_get_address_outputs.rs
@@ -16,7 +16,7 @@ async fn main() -> Result<()> {
 
     let address = "atoi1qzt0nhsf38nh6rs4p6zs5knqp6psgha9wsv74uajqgjmwc75ugupx3y7x0r";
 
-    let outputs = iota.get_address().outputs(&address, Default::default()).await.unwrap();
+    let outputs = iota.get_address().outputs(address, Default::default()).await.unwrap();
 
     println!("The outputs of address {:?} are: {:?}", address, outputs);
     Ok(())